Implementation of the broken line fit
Create modifiers for enabling the broken line fit on the cpu and on the gpu.

Use dinamically-sized-matrices: the advantage over statically-sized ones is
that the code would also work with n>4); the switch can be easily done at the
start of the file.

Update Eigen tests with the features used by the broken line fit.