Do not disable the CUDAService at HLT if there are no GPUs
The CUDAService already checks for the availability of a suitable GPU at job startup,
so there is no need to explicitly disable it in the python configuration.

This fixes the case where the configuration is expanded or pickled on a machine
without GPUs, and then run on a machine with a GPU: in this case the CUDAService gets
incorrectly disabled, breaking the configuration itself.